requirements.txt
playwright==1.47.0
pandas==2.2.2
pytz==2024.1
python-dateutil==2.9.0.post0
